requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
Pillow>=10.4.0
numpy>=1.24,<2
